发票管理修改消息提醒

+++ we-web-parent/web-linyun-airline/src/main/java/com/linyun/airline/admin/user/entity/TSalaryIncreaseEntity.java
@@ -57,7 +57,7 @@ public class TSalaryIncreaseEntity implements Serializable {
 
 	@Column
 	@Comment("提成")
-	private Double commission;
+	private float commission;
 
 	@Column
 	@Comment("罚款")
