+++ we-web-parent/web-linyun-airline/src/main/java/com/linyun/airline/admin/user/form/TSalaryIncreaseUpdateForm.java
@@ -40,7 +40,7 @@ public class TSalaryIncreaseUpdateForm extends ModForm {
 	private Double ratepaying;
 
 	//提成
-	private Double commission;
+	private float commission;
 
 	//罚款
 	private Double forfeit;
